actions

+++ 23-delete-old-logs.sh
@@ -16,9 +16,13 @@ fi
 FILES_TO_DELETE=$(find $LOGS_DIR -name "*.log" -mtime +14)
 # echo "$FILES_TO_DELETE"
 
-while IFS= read -r filepath; do
-  # Process each line here
-  echo "Deleting file: $filepath"
-  rm -f $filepath
-  echo "Deleted file: $filepath"
-done <<< $FILES_TO_DELETE
+if [ -z "$FILES_TO_DELETE" ]; then
+    echo "No files older than 14 days to delete."
+  else
+    while IFS= read -r filepath; do
+      # Process each line here
+      echo "Deleting file: $filepath"
+      rm -f $filepath
+      echo "Deleted file: $filepath"
+    done <<< $FILES_TO_DELETE
+fi
